requirements.txt
django==2.1.7
celery==4.2.1
django-celery-results==1.0.1
django-crispy-forms==1.7.2
simplejson==3.16
opencv-python==3.4.3.18
PyGObject==3.30.1
reportlab==3.5.9
django-guardian==2.1
